+++ project_order/models.py
@@ -36,7 +36,7 @@ class ProjectOrder(models.Model):
     # 定义项目结算模型
 
     project_order = models.OneToOneField(SampleRecord, verbose_name="项目编号", on_delete=models.CASCADE)
-    project_sum = models.PositiveIntegerField(verbose_name="项目金额", null=True, blank=True)
+    project_sum = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="项目金额", null=True, blank=True)
     sale_person = models.CharField(max_length=32, verbose_name="销售人员", null=True, blank=True)
     pay_type = models.CharField(max_length=32, verbose_name="结算方式", null=True, blank=True)
     contract_record = models.BooleanField(verbose_name="合同记录", default=False)
